refactor: clean up diagnostics taxonomy
