+++ crates/diagnostics/src/lib.rs
@@ -2,6 +2,7 @@ mod diagnostic;
 mod result;
 mod sink;
 
+pub mod attribute;
 pub mod emit;
 pub mod infer;
 pub mod lint;
