+++ crates/semantics/src/facts.rs
@@ -2,7 +2,6 @@ use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU32, Ordering};
 
-use diagnostics::lint::MismatchedTailKind;
 use diagnostics::{PatternIssue, UnusedExpressionKind};
 use syntax::ast::{BindingId, BindingKind, DeadCodeCause, Span};
 use syntax::types::Type;
@@ -153,14 +152,12 @@ impl Facts {
     pub fn add_discarded_tail(
         &mut self,
         span: Span,
-        kind: MismatchedTailKind,
         return_type: String,
         expected_span: Span,
         expected_type: String,
     ) {
         self.discarded_tail_expressions.push(DiscardedTailFact {
             span,
-            kind,
             return_type,
             expected_span,
             expected_type,
@@ -305,7 +302,6 @@ pub struct UnusedExpressionFact {
 #[derive(Debug, Clone)]
 pub struct DiscardedTailFact {
     pub span: Span,
-    pub kind: MismatchedTailKind,
     pub return_type: String,
     pub expected_span: Span,
     pub expected_type: String,
